+++ shared/database_base/models/user.py
@@ -1,7 +1,7 @@
 import uuid
 
 from passlib.context import CryptContext
-from sqlalchemy import Column, DateTime, Enum, String
+from sqlalchemy import Boolean, Column, DateTime, Enum, Integer, String
 from sqlalchemy.dialects.postgresql import (  # Although using SQLite, keep for future compatibility
     UUID,
 )
@@ -14,22 +14,16 @@ pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 class User(Base):
     __tablename__ = "users"
 
-    user_id = Column(
-        String(36),
-        primary_key=True,
-        default=lambda: str(uuid.uuid4()),
-        unique=True,
-        nullable=False,
-    )
+    id = Column(Integer, primary_key=True, index=True)
     username = Column(String(80), unique=True, nullable=False, index=True)
     hashed_password = Column(String(256), nullable=False)
+    is_active = Column(Boolean, default=False)
     email = Column(String(120), unique=True, nullable=False, index=True)  # Added email
     first_name = Column(String(80), nullable=True)
     last_name = Column(String(80), nullable=True)
-    role = Column(
-        Enum("user", "super_user", name="user_role"), default="user", nullable=False
-    )
-    refresh_token = Column(String(256), nullable=True)  # Store hashed refresh token
+    roles = Column(String(256), default="user")
+    refresh_token = Column(String(512), nullable=True)  # Store hashed refresh token
+    refresh_token_expires_at = Column(DateTime, nullable=True) # Added new column
     otp_code = Column(String(6), nullable=True)
     otp_expiry = Column(DateTime, nullable=True)